+++ wefact/enums/transaction_actions.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class TransactionAction(str, Enum):
+    MATCH = "match"
+    IGNORE = "ignore"